+++ packages/vue-query/src/__tests__/useMutationState.test.ts
@@ -1,0 +1,80 @@
+import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
+import { ref } from 'vue-demi'
+import { sleep } from '@tanstack/query-test-utils'
+import { useMutation } from '../useMutation'
+import { useMutationState } from '../useMutationState'
+import { useQueryClient } from '../useQueryClient'
+
+vi.mock('../useQueryClient')
+
+describe('useMutationState', () => {
+  beforeEach(() => {
+    vi.useFakeTimers()
+  })
+
+  afterEach(() => {
+    vi.useRealTimers()
+  })
+
+  it('should return variables after calling mutate 1', () => {
+    const mutationKey = ['mutation']
+    const variables = 'foo123'
+
+    const { mutate } = useMutation({
+      mutationKey: mutationKey,
+      mutationFn: (params: string) => sleep(0).then(() => params),
+    })
+
+    mutate(variables)
+
+    const mutationState = useMutationState({
+      filters: { mutationKey, status: 'pending' },
+      select: (mutation) => mutation.state.variables,
+    })
+
+    expect(mutationState.value).toEqual([variables])
+  })
+
+  it('should return variables after calling mutate 2', () => {
+    const queryClient = useQueryClient()
+    queryClient.clear()
+    const mutationKey = ['mutation']
+    const variables = 'bar234'
+
+    const { mutate } = useMutation({
+      mutationKey: mutationKey,
+      mutationFn: (params: string) => sleep(0).then(() => params),
+    })
+
+    mutate(variables)
+
+    const mutationState = useMutationState()
+
+    expect(mutationState.value[0]?.variables).toEqual(variables)
+  })
+
+  it('should work with options getter and be reactive', async () => {
+    const keyRef = ref('useMutationStateGetter2')
+    const variables = 'foo123'
+
+    const { mutate } = useMutation({
+      mutationKey: ['useMutationStateGetter'],
+      mutationFn: (params: string) => sleep(10).then(() => params),
+    })
+
+    mutate(variables)
+
+    const mutationState = useMutationState(() => ({
+      filters: { mutationKey: [keyRef.value], status: 'pending' },
+      select: (mutation) => mutation.state.variables,
+    }))
+
+    expect(mutationState.value).toEqual([])
+
+    keyRef.value = 'useMutationStateGetter'
+
+    await vi.advanceTimersByTimeAsync(0)
+
+    expect(mutationState.value).toEqual([variables])
+  })
+})
